Move domain.py from nodes folder to core because its the only file in that folder.
